Major update and working on a restructuring of code.
Will commit files separately later and do a proper breakdown (Github desktop is playing up and GitHub isn't registering the commits so I have to manually upload the files). There is multiple files as the documentation hasn't been updated properly and i need to put in my identifying and defining and lay my project out really well as it's missing a lot rn.