Update and commit with a code restructure
The documentation needs to have the researching and identifying details put in from the google doc it was completed on but I have updated the code and removed the millions of copies and put them to the documention juypter notebook. I also cleared the hundreds of CSV's made from testing and checking but I have a to-do list and I will release a more detailed update with a almost finalised code and I'll work on documentation.