Solution found!
I managed to use Iloc and used the in place function of pandas to save it. It now completely saves the coin list and I should be able to reuse some of my code for the other function. Looking at the csv saved (before I dropped everything) I could probably get both in the name of optimisation. However its easier to read and I'm alrady committed to it.