The program is completed!
The program is completed and ticks all of the functional criteria. Finally got around to copy and pasting in The documentation but it does need some updates and then I will be completed. Both Viggo and Max have peer reviewed and I just need to double check my code for any errors. As well as fix some slight spelling errors and whatnot +(tick all of the criteria) to gain a high mark. Furthermore the program is receiving an update in the index file and finally getting some infomation. Note that the documentation isn't finished